fix: - refactored the gigantic profile details file into separate files to make it more manageable contextwise
