+++ src/screens/Profile/components/index.ts
@@ -1,0 +1,5 @@
+export { GridGallery } from './GridGallery';
+export { ProfileHeader } from './ProfileHeader';
+export { ProfileStats } from './ProfileStats';
+export { PostModal } from './PostModal';
+export type { RecentDrink } from './GridGallery';
